Remove category list class from Moolah Manager to replace with a hashmap
