+++ src/main/java/seedu/duke/exception/ListStatsMissingTagException.java
@@ -1,0 +1,15 @@
+package seedu.duke.exception;
+
+import seedu.duke.common.ErrorMessages;
+
+public class ListStatisticsMissingTagException extends MoolahException {
+    /**
+     * Returns the error message of the exception to alert user of the exception.
+     *
+     * @return A string containing the error message
+     */
+    @Override
+    public String getMessage() {
+        return ErrorMessages.ERROR_ADD_COMMAND_MISSING_TAG.toString();
+    }
+}
